requirements.txt
Flask==1.0.2
gunicorn==19.9.0
pandas==0.23.4
sklearn==0.0
konlpy==0.5.1
google-cloud-storage==1.13.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-RESTful==0.3.6
tensorflow==1.9.0
numpy==1.15.2